validate notes

+++ backend/Admin/adminnote.js
@@ -59,7 +59,10 @@ router.post("/addnote",async (req,res)=>{
     const token=req.headers.authorization.slice(7)
     console.log(token)
     const data=jwt.verify(token,process.env.JWT_KEY);
-    const newnote=new notemodel(req.body)
+    const newnote=new notemodel({ title:req.body.title,
+            subtitle:req.body.subtitle,
+            category:req.body.category,
+            description:req.body.description,approved:true})
     await newnote.save()
     res.send({message:"successful"})
 })
@@ -69,6 +72,15 @@ router.put("/editnote/:id",async (req,res)=>{
     res.send(value)
 })
 router.delete("/deletenote/:id",async (req,res)=>{
+     if(!req.headers.authorization)
+    {
+         res.status(400).json({message:"error"})
+        return;
+       
+    }
+     const token=req.headers.authorization.slice(7)
+    console.log(token)
+    let data=jwt.verify(token,process.env.JWT_KEY)
     try{
         const del=await notemodel.deleteOne({_id:req.params.id}).exec()
         res.send(del)
@@ -78,6 +90,61 @@ router.delete("/deletenote/:id",async (req,res)=>{
         res.send(err)
     }
 })
+router.get("/approve",async (req,res)=>{
+     if(!req.headers.authorization)
+    {
+         res.status(400).json({message:"error"})
+        return;
+       
+    }
+     const token=req.headers.authorization.slice(7)
+    console.log(token)
+    let data=jwt.verify(token,process.env.JWT_KEY)
+    let value=await notemodel.find({approved:true}).exec()
+    res.send(value)
+})
+router.get("/reject",async (req,res)=>{
+     if(!req.headers.authorization)
+    {
+         res.status(400).json({message:"error"})
+        return;
+       
+    }
+     const token=req.headers.authorization.slice(7)
+    console.log(token)
+    let data=jwt.verify(token,process.env.JWT_KEY)
+    let value=await notemodel.find({approved:false}).exec()
+    res.send(value)
+})
+router.put("/approve/:id",async (req,res)=>{
+     const cid=req.params.id;
+    let value=await notemodel.findByIdAndUpdate(cid,{approved:true},{new:true})
+    res.send(value)
+})
+router.delete("/delete/reject/:id",async (req,res)=>{
+     if(!req.headers.authorization)
+    {
+         res.status(400).json({message:"error"})
+        return;
+       
+    }
+     const token=req.headers.authorization.slice(7)
+    console.log(token)
+    let data=jwt.verify(token,process.env.JWT_KEY)
+    const value=await notemodel.findById({_id:req.params.id})
+    if(value.approved==false)
+    {
+        try{
+        const del=await notemodel.deleteOne({_id:req.params.id}).exec()
+        res.send(del)
+    }
+    catch(err)
+    {
+        res.send(err)
+    }
+    }
+     
+    
 
-
+})
 module.exports=router;
